Update patch set 9
Patch Set 9:

(2 comments)

Happy to discuss the details. I was equally as baffled before I started working on this.

In PS10, I also removed the 'cache_position_' field which I had used in debug build to ensure the ordering was correct. Now the DCHECKs only verify that the search doesn't miss any results, which is less strict but sufficient.

Patch-set: 9
Label: Verified=0