Update patch set 13
Patch Set 13: Code-Review+1

> I don't understand this change:
This changes tries to work on the premise that if the application is not debuggable and if either the caller or callee do not have catch then the callee context is done on exception throw.

 > 1) References to try/catch that optimizing does not handle yet.
This is only to explicitly capture the premise on which this works. You are right, that catches are not allowed by optimizing in either caller or callee. When that support is eventually added, we still want this code to work and hence why the checks for catch blocks are explicitly added in inliner. Another approach is to simply assert it but I found it cleaner implementing it this way.

 > 2) How can that work with stack traces?
It does not work with stack traces actually and hence why I explicitly check that graph is not debuggable. If we want to make it work with stack traces, we can update runtime to better determine location that is throwing by taking a look not only at dex offset, but also at inline info to recover stack.

Patch-set: 13
Label: Code-Review=+1