Update patch set 2
Patch Set 2:

> Rebased.
 > 
 > I am still baffled by the exact sequence needed to remove the
 > compare and cond, and replace the If with a CompareIf.
 > 
 > If I replace the If with a CompareIf, then calling
 > RemoveInstruction on the cond and the compare either dies with
 > art/compiler/optimizing/nodes.cc:503] Check failed:
 > instruction->GetEnvUses().IsEmpty() when removing the compare after
 > I removed the cond.
 > 
 > If I remove the compare first, it dies with art/compiler/optimizing/nodes.cc:502]
 > Check failed: instruction->GetUses().IsEmpty() while removing the
 > compare.
 > 
 > What am I missing here?  It makes sense to remove the Cond first,
 > as that will remove the use of the Compare.  What needs to be done
 > with the env_uses on the compare?

OK, I see what the problem is. You need a liveness analysis to figure out that the HCompare is only used once. Today, the DEX register for the DEX compare instruction gets the -1/0/1 value and every instruction that requires an environment will hold it. Only after liveness we end up knowing it's only used once.

I suggest punting this change for now, we'll get back to it when we have better infrastructure for doing this.

Patch-set: 2