Update patch set 3
Patch Set 3:

I think I could do another implementation, where I set a flag in the compare that it doesn't need to be manifested if it is only used for this if.  I would then handle the logic in the VisitIf code generation.  Would that work?  Would the environment be okay if the vreg wasn't created?

Patch-set: 3