Update patch set 3
Patch Set 3:

> How about if debugging is turned off?

I meant debugging/deopting, where deopting should always be turned on. You'd need to make sure 1) there is no deopt point between the HCompare and the HEquals and 2) the HEquals is the only user.

Before rushing anything, we need to get 1) right, so that it's not just for this change, but for others. I guess in the meantime, you could just check that you have HCompare/HEqual/HIf in order, and without any instruction in-between.

Patch-set: 3
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0