Update patch set 1
Patch Set 1:

x86_64/core.oat.x:
- 11 .oat_patches  00082878  00000000  00000000  01b42f28  2**3
+ 11 .oat_patches  000146ea  00000000  00000000  01b42f28  2**0

arm/core.oat.x:
- 11 .oat_patches  00066a80  00000000  00000000  0186ce00  2**3
+ 11 .oat_patches  0000fb93  00000000  00000000  0186cdf9  2**0

Oh, I just realized there is catch.  It is because I am running this on host - so it is 64-bit compiler.  I expected that if I run 32-bit compiler, the old number would half.  So it would be just factor of 3.

The above number are without an DWARF data.

Patch-set: 1