Update patch set 2
Patch Set 2:

> Just use ATRACE_TAG_DALVIK for these

A lot of other information is logged to ATRACE_TAG_DALVIK that is not relevant for understanding performance of boot. If I don't have a separate ATRACE_TAG_BOOT, then the boot trace I collect will be much larger and noisier than I want.

Is there some other approach in systrace to filter out events to reduce the trace size and noise?

Patch-set: 2