Update patch set 5
Patch Set 5:

(11 comments)

Actually, it is scratch location. The NoSwap implementation doesn't care about what the location really is. It is the backend's responsibility to allocate a suitable scratch location. It can be used to fix the 32-bit swap issue.

For example, moves (R0->stack) (stack->R1) (R1->R0) will be performed as (R1->scratch) (stack->R1) (R0->stack) (scratch->R0).

Here R1->R0 might be a 64-bit or 32-bit move, and it is the backend choice what the scratch location should be. In the ARM64 implementation in this patch, the scratch location will be a register(64-bit). Because I don't know the size of R1 and R0. The scratch location is used to save R1 temporarily.

But there are some tricky things for intersecting cycles. If this is the case for x86_64, I prefer to implement arm32 backend first and take that one as an example for x86_64.

Patch-set: 5
Label: Verified=0