Update patch set 1
Patch Set 1: Code-Review+1

Lgtm. I agree with Zheng Xu that type information in location seems better in general case but I bet it is a big re-factoring, probably it might be done later. Also as soon as we know the type of the move, may be it makes sense to transfer this information to EmitMove and EmitSwap, so they could generate better code?

Patch-set: 1
Label: Code-Review=+1