Update patch set 1
Patch Set 1:

I used the dex file to run the Logic benchmark, and the problems seems to be replacing the
   push reg; mov reg,src; mov dst, reg; pop reg
with
   push src; pop dst

Undoing that seems to have returned the performance.

I started playing around undoing other changes, trying to make sure that the speedups where all good, but the variation in the runtime seems to be greater than the performance differences on my host, so I can't reach any conclusions.

I THINK that using a scratch register to do a 3 instruction swap is better than xchgl reg,reg2, but I can't really confirm that due to the above.   I think that this might just have to remain reverted until I can get our performance guys to give me a final answer.

Patch-set: 1