Update patch set 1
Patch Set 1:

This rough fix of optimizing inliner is proposed until better job is don on graph to eliminate redundant clinit checks. Note that test 473 is about to incorrect. Compiler cannot be sure that class is initialized until first call to that class is done. So I propose to make the negative case to be the first invocation. so this call will ensure that class is initialized.

Patch-set: 1
Reviewer: Gerrit User 1040038 <1040038@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0