Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > BTW, do you think this change could be exercised with a test? 
 > Maybe that's the kind of things we would like to check with
 > Checker, when it supports assembly language checking?

I am not sure how a test could help here. The issue is that to test that no code is inserted in locations where we care about the PC, we need to identify these locations.
Another approach is to use VIXL helpers to check that the code 'just before' we use the PC is what we expect.

So for a pseudo-code example for CFI we coud do:

    __ Str(x0, MemOperand(sp, offset));
    CFIDoStuffWithPC();

    void CFIDoStufWithPC() {
    #ifndef NDEBUG
      vixl::Instruction* prev_instr = vixl_asm->InstructionAt(pc - kInstructionSize);
      DCHECK(prev_instr->IsStore() && prev_instr->Rd() == 0 && prev_instr->Rn() == ...);
    #endif

      // Do stuff.
    }

Patch-set: 2