Update patch set 3
Patch Set 3:

> IIUC, you force the default locale to be ENGLISH when initialize
 > parts of tagsoup ahead-of-time. Would the initialization path be
 > equivalent if another locale was the default (according to
 > user.locale system property)? Would it work with a default language
 > like Japanese for instance?

HTML tags are ASCII. Lower-casing ASCII characters is well-defined and should not change wrt/ locales.

 > why is this in the runtime ? it's probably less icky (and correct)
 > to change uses of String.toLowerCase() in tagsoup to
 > String.toLowerCase(Locale.ROOT). otherwise they'll end up with
 > surprising tags in languages like turkish.

They won't end up with surprising tags, as HTML is well-defined. Tags are ASCII. Otherwise I would not have done this. :-)

I put it into the runtime as tagsoup is external. If we all agree to change it in external/ or upstream I'd be very happy to remove this.

Note that we already do certain things based on implementation knowledge, e.g., fake ThreadLocal support for number conversion.

Patch-set: 3