Update patch set 1
Patch Set 1:

> > > Does this work correctly wrt/ inlining?
 > >
 > > I don't see why there would be any problems with inlining.
 > >
 > > 32-bit and 64-bit boot.oat on Nexus 9 are reduced by 28KiB and 52
 > > KiB, respectively.
 > 
 > What's at the bottom of the stack frame of the caller when we
 > inline? I believe that question wasn't resolved, yet. I believe
 > it's working right now because optimizing currently cancels
 > inlining when the dex cache might be needed and it's not the same
 > as of the "caller." But in general you're making a change here now
 > that we discussed before and postponed (else you *should* have
 > dusted off my WIP).

Yeah, we discussed about this during that prior CL, and even though we're safe today, assuming that the current method is at the bottom of the stack could easily break once we do more inlining.

If these calls are rare, why not recovering the method from the StackVisitor?

Patch-set: 1
Label: Code-Review=0