Update patch set 1
Patch Set 1:

Currently, the compiler knows it needs to load the ArtMethod* and the entrypoint needs to forward it. With this change, the compiler is no longer concerned, only the entrypoint loads it. So there will actually fewer places that need to be modified for inlining.

If we want to start inlining methods with these calls, we shall have to use the StackVisitor anyway. Does anyone know approximately how expensive it is or do I need to benchmark myself? (To decide whether we want to use StackVisitor now or wait until we actually inline.)

Patch-set: 1
Label: Verified=0