Update patch set 1
Patch Set 1:

Oh, I did not realize there was similar code in assembler_test.h.  Sorry!

I have found the version in assembler_thumb_test when I was doing dwarf tests, but it was Thumb2 specific and I needed host's prebuilds so I copy-pasted it and adjusted it.  I felt bad for the copy-paste so I tried to unify the two now and put them somewhere where they can be access by the dwarf tests as well as assembly tests.

Patch-set: 1