Update patch set 1
Patch Set 1:

Maybe the right solution is to have a big and a little set. Together you will get the big.litttle combo. However we would need to have a list of possible combinations. I.e a cortex-a57.cortex-a7 is a invalid option. Though cortex-a57 is a valid big and cortex-a7 a valid little.

I will think about it and take a look at what GCC does. In the meantime can you take a look at the build changes that I have proposed? This change very much depends of the build changes added in in the attached patch.

Patch-set: 1