Update patch set 10
Patch Set 10:

This patch addresses all the concerns expressed earlier. For the host, all ISAs are supported. For target, only TARGET_ARCH is supported.

Please compare against base as this change is very different from previous versions of the patch.

Concern about removing -Wredundant-decls : The reason for removing it is because, the dummy files in error folder implement a dummy function that returns nullptr.

Concern about TARGET_2ND_ARCH - In case of 64 bit TARGET_ARCH, the 32 bit is already supported. For example, if TARGET_ARCH is x86_64 and TARGET_2ND_ARCH is x86, x86_64 already supports x86.

Patch-set: 10
Label: Verified=0