Update patch set 1
Patch Set 1:

> Wait, you completely remove the capability of setting values with
 > debugger: how does it not break related JDWP tests ??? Can you
 > explain why you need to do that?

The JDWP tests request the --debuggable flag, which lead to using baseline in optimziing. baseline has the same stack layout as Quick, so for it, we're using the Quick routines.

 > 
 > The alternative way is to deoptimize the stack to set values in a
 > frame. However the support is not done yet. We should have dropped
 > this code only after that step.

I removed it to avoid giving bogus execution while being debugged. It would be kind of surprising to update local 1, but because of SSA, local 2 and 3 are updated to the same value :-)

Patch-set: 1