Update patch set 3
Patch Set 3:

I can't submit this CL yet, as it depends on https://android-review.googlesource.com/#/c/148046/, which is not yet in.  Guillaume, if this doesn't break the build, you may want to reverse the order of these commits and re-upload them, so that this one can be merged before the other.

Patch-set: 3