Update patch set 2
Patch Set 2:

(106 comments)

> (131 comments)
 > 
 > Thanks for this huge work!  As requested by Andreas, could you
 > please extend codegen_test.cc?

Can the UI glitches be fixed? The file review page with the comments is always scrolling away from me. Whenever I click on a comment or just type a response or even when I merely scroll the page using the mouse wheel, I suddenly appear in a different place on the page, typically several screens above or below the place where I need to do something. This is happening on FireFox/Ubuntu14x64 and Chrome/Win7x64 and is extremely frustrating for the number of the comments to take care of. Any help?

Patch-set: 2
Label: Verified=0