Update patch set 17
Patch Set 17:

> is this up for reviewing or still experimenting?

Oh right, sorry for the lack of comments.  PS17 is a rebase (with many changes, as the code has changed quite a lot since PS16).

I think this CL could be reviewed as-is, but I'd like to add some more changes to it.  So, could you wait for my signal before reviewing it?  Thanks in advance. :)

Patch-set: 17
Reviewer: Gerrit User 1052644 <1052644@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0