Update patch set 1
Patch Set 1:

(1 comment)

I think I probably copied some code earlier.  I was just assuming that since I know that the invoke has a hardcoded EAX in it, that it would be safest to ensure that EAX was always an input or temp.   I suspect that you are correct that it is irrelevant, but I know from many years experience that it is very bad to lie to an optimizing compiler :-)

Nicolas would know, I expect.

Patch-set: 1
Label: Code-Review=0