Update patch set 5
Patch Set 5:

Rebased.

I tried to redo this to create the instruction while generating locations, but it doesn't work because the new instruction didn't have a lifetime allocated to it.  I wonder if a gap can be left for the BEs at the end of the first BB?

Patch-set: 5
Label: Verified=0