Update patch set 1
Patch Set 1:

Just to make sure we are talking about the same thing. The core image is not the boot image. The core image is only used for our (the ART team) testing. We have two kinds of tests: gtest and run-test. The run-tests do not need junit. I ran them, they pass.

My intuition is that it was added for convenience, and was harmless because it's easy to compile and the boot image also has it.

Patch-set: 1