Update patch set 1
Patch Set 1:

> Serban,
 > 
 > Even though the idea sounds interesting, this doesn't help archs
 > that have callee-save registers at the end of the registers array
 > (so arm32, x86 and arm64). Linear scan does already the job for:
 > 
 > 1) Intervals that do not span a safepoint: it does not intersect
 > with blocked registers due to calls, so it will pick the first
 > fully available. This CL only helps on x64, as there are
 > caller-save registers at the end of the array that could be
 > available.
 > 
 > 2) Intervals that do span a safepoint: caller-saved registers will
 > be blocked, so free callee-save registers will always be available
 > longer than caller-save.
 > 
 > It looks like you investigated quite some time for this on arm64,
 > so I may be missing something. Please let me know.

Need to amend this a bit: it only helps archs that do not have a [caller_save, callee_save] pattern for its registers. Only x64 has this, with [caller_save, callee_save, caller_save, callee_save].

So this saves some moves on x64 in the case where a caller-save interval takes a register in the first callee-save range, and a callee-save interval only has caller-save availables. Should be pretty rare.

The main advantage I see is less callee-save saved at entry of a frame because now a caller-save interval will reach the caller-save of the second range.

Patch-set: 1