Update patch set 9
Patch Set 9:

Rewritten to use prepare_for_register_allocation.cc rather than the instruction simplifier.

Since we know that there is no intervening instruction, it would be nice to have some way to remove the HCompare from the environment of subsequent calls, since this isn't debuggable.

Does it make sense to remove the restriction that these are consecutive instructions, and just check to see that neither the condition nor the compare is referenced in an environment?

Patch-set: 9
Label: Verified=0