Update patch set 10
Patch Set 10:

Uploaded a version that removes the HCompare in prepare_for_register_allocation, moving the information into the HCondition instead.  Has versions that support x86 and x86_64 only at this time.

Also contains code that, if a #define FAILS_COMPILING_CORE_OAT is added, will cause an 'out of registers' crash compiling core.oat for x86.  This is because I ask for 2 temporary registers for a long HCondition, and that seems to be too much.

Question:  Is there a way to see if an input goes dead in an instruction?  Is asking if there is only 1 use of the input enough?  I would like to use that information to remove the need to copy a register's contents to the temporary, thus saving some useless instructions.

Patch-set: 10
Label: Verified=0