Update patch set 7
Patch Set 7:

Let me know if there is anything that I can do to help with the arch-specific optimizations.  I thought that this scheme was fairly general.

Note that this change does allow much better FP code gen in x86 and can be extended for more improvements.

Patch-set: 7
Label: Verified=0