Update patch set 1
Patch Set 1:

I am sorry, I do not follow.  Which flag do you mean?

My point is the at the moment function names can be generated only by passing --include-debug-symbols to compiler (which is default in debug build).  So we can get CFI+names only be generating everything and then stripping the unnecessary DWARF sections (which are not cheap to generate in the first place).

I would rather be in situation where we can easily generate CFI+names (and strip names if necessary) without needed to waste effort on the full DWARF sections.

Patch-set: 1