Update patch set 2
Patch Set 2:

I just skip around deoptimizations. No skip around parallel moves. The jump is to the instruction following the last HDeoptimize.

Since deopt can hoist some values (like array.length after a null check deopt, which we want to skip), this is really handled specially. We kind of violate the SSA form so I don't see a good way to solve it at a higher level since if it jumps around, the array.length value doesn't exist. But the value can't be there without a null check deopt, which we want to skip. In the way I did we know jumping around and going to the loop header without the array.length value is safe since we know the loop isn't entered. I have some handling and comments in CanAddDeoptimizationArrayLength() about it.

 > How can it be safe to skip around parallel moves? We might miss
 > some that should be executed before entering the loop header.
 > 
 > Also, it bugs me a bit that we need to now handle it in the code
 > generators. At this point, isn't it cleaner/easier to just add a
 > block before the loop header?
 > 
 > 
 > I think you can easily add an extra block before the loop header,
 > that contains all the deopt and the new compare instruction.

 > How can it be safe to skip around parallel moves? We might miss
 > some that should be executed before entering the loop header.
 > 
 > Also, it bugs me a bit that we need to now handle it in the code
 > generators. At this point, isn't it cleaner/easier to just add a
 > block before the loop header?
 > 
 > 
 > I think you can easily add an extra block before the loop header,
 > that contains all the deopt and the new compare instruction.

Patch-set: 2