Update patch set 2
Patch Set 2:

Well that CL makes sense - it enforces split of two function which use a lot of stack.
In this case the nested calls take barely any stack space.  The only way they could significantly increase the stack size of the main function would be if the nested calls generated many temporary variables on the stack.

Patch-set: 2