Update patch set 7
Patch Set 7:

> Thanks Alexey!  I'll test your changes ASAP.

I used a patch that is currently not on AOSP, AFAIK, (expecting it to be uploaded later today or next week), so your results may be not as good as mine (Android boots here to home screen and most of ART tests pass). For this reason I haven't commented on exactly what works and and what doesn't. Sorry. Distributed work at work.

Patch-set: 7
Label: Verified=0