Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

There are multiple places calling DumpNativeStack. The design with a global that is only maintained in one of the callers and not protected by a lock seems inherently racy. Either there need to be at least DEBUG checks that there are no mixed uses, or you should refactor this so that the caller passes a map down if it cares.

Patch-set: 1
Label: Code-Review=-1