Update patch set 3
Patch Set 3:

I am actually somewhat surprised that is works.
The String.valueOf method is pretty trivial.
We can ensure that foo and bar are not inlinable, but I think that would not work since it is valueOf that we want on the stack.
I am thinking of Arrays.binarySearch now since it has loop inside.

Patch-set: 3