Update patch set 3
Patch Set 3:

On second thought, the code of valueOf is:
  return value != null ? value.toString() : "null";
which is trivial, but I think the branch should be sufficient to prevent inclining?

valueOf is kind of neat because it is obvious it calls bar precisely once, and it easily encodes the return value.  I think that Arrays.binarySearch can achieve the same thing, but it is less neat.  What do you think?

Patch-set: 3
Label: Verified=0