Update patch set 1
Patch Set 1:

I agree with that.  I do not like dead code either.  If we had complete control of the input dex file I would happily crash on it.  However, the dex file comes from outside and it can be generated by any version of java compiler - potentially future ones, and potentially intentionally modified to be more debugger friendly.  Can the dex file be generated other tools than java compiler?  (say, the dex file has support for having different filename for each bytecode, but java language do not have support for that at all, so that is where I drew the line for the time being and I do not support that - where "not support" means it will not crash, it will just ignore any mid-function filename updates)

Long story short, I would call this a bit defensive coding.

Small until test would be nice (especially for mid-function filename changes, and potentially other tests of the debug-info which is not supported yet) but I do not have neither the time nor knowledge how to generate the custom dex file.  I am not too worried about the code as it is - the body of the "if" statement is the path that will be taken now, so condition will just skip that mangling.

Patch-set: 1