Update patch set 11
Patch Set 11:

> > I rebased this CL on the top of the tree.  dex2oat fails to
 > compile
 > > /system/app/BasicDreams/BasicDreams.apk because of failures in
 > the
 > > register allocator:
 > 
 > I can fix this failure in dex2oat by actually using the is_baseline
 > parameter in CodeGeneratorMIPS64::SetupBlockedRegisters(), but we
 > still seem to have clobbering of MIPS64 S- (saved) registers
 > somewhere else, so, the generated code isn't functional with this
 > "fix".
 > 
 > If I understand the latest changes correctly, the failure is
 > triggered by the new AddTemp(visitor->GetMethodLocation()) calls in
 > CodeGenerator::CreateCommonInvokeLocationSummary(), which increase
 > the pressure on the available registers. But maybe that
 > understanding is incorrect. I wonder how we manage to run out of so
 > many MIPS64 V-, A- and T- registers (2+8+4=14) that aren't blocked
 > in CodeGeneratorMIPS64::SetupBlockedRegisters() and need to
 > additionally grab an S- register, whereas there are many fewer
 > registers available on x86 and x86 is somehow OK with that. All
 > this logic in the register allocator and the parallel mover and
 > such should efficiently spill registers onto stack and resolve the
 > register pressure, right?
 > 
 > I'm probably missing something here.
 > Any help?

You'll get this failure by requesting too many registers for an instruction: here it's an HInvokeStaticOrDirect. We hat tat failure ourselves for x86 and baseline because we did not have enough registers for it. The "fix" was to simply not require a register the the current method, but do the load on the stack ourselves.

If anything, you can try do the same logic to fix it. But I agree with you that you should not get that error with that many registers. Or is the ABI for calling in mips64 using all core registers for parameters?

Patch-set: 11