Update patch set 1
Patch Set 1:

Initially we used signal chaining just for implicit checks but now we also use it for the native bridge. 

Now I know that actually that code path is not called during dex2oat so that it's not necessary a blocker (given that we add some assertion to make this clear).

The good part of having something like is_dex2oat is that we could potentially disable more from the runtime with just a single flag. But I'm not sure yet how beneficial would that be.

Patch-set: 1
Label: Verified=0