Update patch set 16
Patch Set 16:

Your suggestion is a pretty good fit for the basic math operations.  The HLoadConstantFromConstantPool would also need a HConstant input as well, so we know which one to generate.

There are other instructions that would like to use the constant area, and would work fine with this scheme:
- compares
- field sets

There are other instructions that would like to use the constant area that don't fit as nicely on the X86:
- FP neg on x86 (for the constant xor mask)
- packed switch (to address the jump table)
- FP intrinsics (max/min/abs) (for masks/constants)

We also want to address the constant area for:
- invokes (to load FP values if needed)
- returns (ditto)

The loads for these are generated by HParallelMove (not the BE code gen), which can look to see if the next instruction has a 'backend input', and use that if available (this removes most uses of the stack for loading FP).  For this to work, we need to add a 'backend input' to invokes/returns.  If we replace a FP parameter by a HLCFCP version as an alternate implementation, we want to be able to load directly into the target register.  I don't think RA will be able to do this without bigger changes.

Using the 'backend input' simplifies most of the uses of the above.  Otherwise, the x86 BE will have to create new X86 versions of (e.g. HNeg to add the extra input) or rewrite the implementation of HInvoke to figure out if a parameter is a HLCFCP or an integer value.

Synopsis:  Your suggestion is great for the original uses, but not for more complete use of the constant area in the X86.  It can be made to work as is by adding X86 versions of all of the above instructions, and doing the rewrite in the X86 Backend optimization pass.  This at least minimizes the changes to other architectures.

Somewhat related: It would also be nice to move arch specific instructions into a nodes_<arch>.h file, but they would have to be #included in the middle of nodes.h in order to get the right macros defined and ensure the right classes are visible.

Comments on my comments?

Patch-set: 16
Label: Verified=0