Update patch set 16
Patch Set 16:

Examples of needing inputs that aren't standard inputs:
- HNeg fp wants both the fp input and also the constant for an xor mask.  We want to have HNegX86 fp, HLCFCP(0x800...).  We also need a temp location to load the value loaded from the constant area, but that is trivial.
- HPackedSwitch needs have the switch value, as well as the base of the constant area (HX86PackedSwitch sw_value HComputeBaseMethodAddress)
- min/max/abs also wants constants from the constant area.

The idea was that adding this backend input is transparent to the register allocator.

Invokes and returns are actually requirements on ParallelMoves as noted.  I am certainly not qualified to add the needed hints to the RA as to how HLCFCP operands can be properly loaded to the right place.  What could work is having EmitMove generate a move from a HLCFCP instruction (rather than the constant itself), but the parallel move generator wouldn't know how to address the base of the constant area without some external help.

I don't want to generate any code when visiting a HCLFCP node, as that will involve loading into a register, which is probably not what I want to generate.  I want to be able to detect uses of the constant area while generating code for (certain) other instructions.  The classic example of this is HAdd, which wants to add directly from the constant area.

Patch-set: 16