Update patch set 16
Patch Set 16:

> I am still expecting that this type of use would be BE specific and
 > only legal to generate in the BE optimization pass. It is meant to
 > overcome BE code generation problems. I am not certain that GVN et
 > al would want to work with this at all.

For example of where GVN might be interesting, we could share the load of a constant, if that load can be in register. So I don't think we should view these new nodes as BE special.

 > 
 > If the uses were restricted to BEs, then deriving HLCFCP from
 > IsConstant is doable.  Another hack would be to just override
 > IsConstant() in HLCFCP.  The idea is that RA would treat it as a
 > constant and just leave it alone.

Why do you need a hack? We should just code it differently. A virtual HasOwnSpillSlot would do.

Patch-set: 16