Update patch set 16
Patch Set 16:

For the X86, I don't want the value commoned and loaded into a register.  I want it addressed directly from memory, as registers are the scarce resource on that architecture.  The commoning that I care about happens in the constant area.

Would you expect to generate a HLCFCP in the builder?  If not, I am not sure why you expect other passes to want to handle it.  If there is a need for commoning, etc after BE optimization, I would expect that BE to invoke other passes as needed.

I assume that HasOwnSpillSlot would be an enhancement to HInstruction/RA to state that it wasn't necessary to spill (or in fact can be recreated easily)? returns true for IsConstant and IsLCFCP?

Patch-set: 16