Update patch set 1
Patch Set 1: Verified+1

We don't have automatic tests for these, which is unfortunate. Mathieu's GC test is stressing this now, as he's using the libc backtrace() function to unwind.

A problem is that GDB is often good enough to unwind automagically even when things are incorrectly annotated. Otherwise you'd try to set a breakpoint on the C side and wait...

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1