Update patch set 1
Patch Set 1:

How hard would it be for ART to execute the shutdown hooks when it terminates "normally"? Seems a shame to have the APIs but not have them work for standalone cases. (Even though Android *framework apps* are warned in the docs this mechanism doesn't work).

It looks like ART has the correct logic for handling shutdown hooks if code explicitly calls System.exit() (actually Runtime.exit()), so I'm assuming that if it isn't invoking the shutdown hooks when the last non-daemon thread dies then it might be relatively straightforward to fix...?

Patch-set: 1