Update patch set 1
Patch Set 1: Code-Review+2

Spoke to Andreas, and I agree that we probably don't want to be setting the register type to conflict even in the soft failure case. This is fine for now to fix the hard failure case though.

Patch-set: 1
Reviewer: Gerrit User 1010108 <1010108@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2