Update patch set 1
Patch Set 1:

Let me know if this is not the right place to add the flag.

http://reviews.llvm.org/D7633 added this warning.  For line 261 in art/runtime/base/variant_map.h, I think there is no need to call std::move on the intermediate object created here.  I am not sure how to test if I remove the std::move.  I'll let someone from ART handle this.

I won't submit the rebase for a few days.  In the mean time, if the call to std::move can be removed, please feel free to submit a CL doing the same.  I'll abandon this CL.

Patch-set: 1
Label: Verified=0