Update patch set 2
Patch Set 2:

(4 comments)

Thanks Roland and David, PTAL.

David: how would you like this tested? With a checker test? It has to be arch-specific, do we support that already?

If we do, do you want to test that the suspend check gets a register for a simple loop? But not if it's a more complex loop?

Patch-set: 2
Label: Verified=0