Update patch set 4
Patch Set 4:

Thanks for taking a look at this, Roland!
Yes, lets wait until I sort out the metadata size. As mentioned in the commit message, the current scheme for encoding the native_lr is not ideal. I am currently refactoring the StackMaps code and implementing a new method for encoding the native_lr. Once that is done I will extend this to all the ARM64 slowpaths and re-run the stats.
I will push an updated version of this patch towards the end of next week when I am back from holiday.

Patch-set: 4