Update patch set 1
Patch Set 1:

> > > Thanks Tony. This seems to fix it rather than refining, right?
 > > >
 > > > Andreas, you reviewed the the code that added it here:
 > > > https://android-review.googlesource.com/#/c/102772. Do you know
 > > > what was the intention? Do we agree the code is just broken?
 > >
 > > This isn't broken per se. It just has a weak, non-intuitive
 > canary
 > > value as-is.
 > 
 > Hmm, maybe you can help me understand what is going on. What values
 > can 'j' hold in the loops? (in the code before this change).

You *still* get a canary. It's the first letter of the canary text repeated, and this will be correctly checked. Technically it doesn't matter what your canary characters are, there's always the chance that a modification will "replace" the character with the same value.

I'm not debating the fix here. I'm stating that red zones are *not* broken.

Patch-set: 1