Update patch set 3
Patch Set 3: Code-Review+2

I'm giving a +2, but it's for you to decide whether you want to merge this as-is or do what I sketch in the next paragraph.

Check art/build/Android.common_build.mk on how we usually add the define to the build flags locally, translating a make variable to a cflags option. Potentially you may want to set this up in a similar way, so as not to pollute the general CFLAGS in the build system.

Patch-set: 3
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2