Update patch set 1
Patch Set 1:

Request for comment:

I wanted to remove the extern to get actual checks in place. Then I wanted to move the enum. That led to the problem of the enum being used, as we want to cache the result for a whole class in the main case - but it should be hidden IMHO in the compiler-driver h file.

Options were forward-declared enum (iam used one before), or this friend class helper to access some internals I did not want to expose with public methods.

Better options? Or is this just overkill?

Patch-set: 1
Label: Verified=0