Update patch set 1
Patch Set 1:

> Andreas, could you please provide some details or ask them to
 > update a bug if possible because I'm not sure I understand what you
 > mean :)

The package manager will hold a wake lock now, which means the original bug report is solved.

A timeout *may* be acceptable. Still unhappy that it touches quite a bit of runtime internals, which could hang anyways (e.g., the FATALing out on timeout will try to dump threads). When I was thinking about this before (to regain dumpability on timeout), the watchdog would have to have three states, with another timeout maybe 1m later that kills everything with an exit immediately.

Patch-set: 1