Update patch set 1
Patch Set 1:

> Andreas - what happens when dex2oat is invoked from class_linker
 > instead of via PackageManager/installd path. Who will hold wakelock
 > in that situation ? PM is not involved there.
 > [e.g. art::ClassLinker::GenerateOatFile execs dex2oat. ]

Sorry, but I do *not* worry about this case. The code is complicated enough as-is. If your app is still behind times and using your own multidex scheme, it's up to you to keep your app alive if you have a dex file that's large enough to trigger this (I also don't worry about custom vendor optimization passes that increase compile times - that's your issue).

(In general, if it's an actual app, I'd assume you're in the foreground and active, so the device shouldn't go to sleep. If it's a service, then you better make sure it's a small dex file that will compile quickly.)

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0